- Finally really fixed not tests - Now passing 1585/1833 (on my machine)
git-svn-id: svn://localhost/mosa@133 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Test/Mosa/Runtime/CompilerFramework/IL/Not.cs
@@ -52,7 +52,7 @@ namespace Test.Mosa.Runtime.CompilerFramework.IL
         public void NotI1(sbyte a)
         {
             CodeSource = "static class Test { static bool NotI1(int expect, sbyte a) { return expect == (~a); } }";
-            Assert.IsTrue((bool)Run<I4_I1>("", "Test", "NotI1", ~a, a));
+            Assert.IsTrue((bool)Run<I4_I1>("", "Test", "NotI1", (sbyte)~a, (sbyte)a));
         }
     }
 }
